% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_nightly.R
\name{use_nightly}
\alias{use_nightly}
\title{Enables the nightly javascript library}
\usage{
use_nightly()
}
\value{
called for the side effect of activating the scrollrevealjs library
}
\description{
The function activates the capabilities of the nightly javascript library. The user can put it anywhere within the UI.
}
\examples{
\donttest{

# Put the function within your shiny ui

use_nightly()



}
}
